delete QuizApp and ApplicationComponent
